requirements.txt
feedparser==6.0.11
newspaper3k==0.2.8
nltk==3.8.1
lxml_html_clean==6.0.0